Remove testing.py
