+++ numba_dppy/tests/njit_tests/dpnp/test_numpy_rng.py
@@ -20,7 +20,7 @@ import dpctl
 import numpy as np
 from numba import njit
 import pytest
-from numba_dppy.testing import dpnp_debug
+from numba_dppy.tests._helper import dpnp_debug
 from .dpnp_skip_test import dpnp_skip_test as skip_test
 from ._helper import wrapper_function
