+++ numba_dppy/tests/njit_tests/dpnp/test_numpy_transcendentals.py
@@ -20,9 +20,8 @@ import dpctl
 import numpy as np
 from numba import njit
 import pytest
-from numba_dppy.testing import dpnp_debug
 from .dpnp_skip_test import dpnp_skip_test as skip_test
-from numba_dppy.tests._helper import is_gen12
+from numba_dppy.tests._helper import is_gen12, dpnp_debug
 from ._helper import wrapper_function
