+++ numba_dppy/tests/test_dpnp_functions.py
@@ -18,7 +18,7 @@ from numba import njit
 import dpctl
 from . import _helper
 import unittest
-from numba_dppy.testing import ensure_dpnp, assert_auto_offloading
+from numba_dppy.tests._helper import ensure_dpnp, assert_auto_offloading, dpnp_debug
 
 
 @unittest.skipUnless(
@@ -39,7 +39,7 @@ class Testdpnp_functions(unittest.TestCase):
             d = c + e
             return d
 
-        with dpctl.device_context("opencl:gpu"), assert_auto_offloading():
+        with dpctl.device_context("opencl:gpu"), assert_auto_offloading(), dpnp_debug():
             njit_f = njit(f)
             got = njit_f(self.a, self.b)
         expected = f(self.a, self.b)
